pyproject.toml
[project]
name = "openaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit",
"openai-agents[sqlite]>=0.2.6",
    "python-decouple>=3.8",
]
